requirements.txt
RPi.GPIO==0.6.5
telepot==12.7

